Remove ability to specify rectangle using negative co-ords.
It prevented the ability to specify rectangles which lie
outside the visible terminal area (which is perfectly possible
with circles - even though both won't render correctly yet
until we add clipping)

It was also one of the two reasons for needing to inject
terminal into the Shape class. The other reason is also
removeable (by extracting the str method to some other module
that knows about 'terminal'). Then the shape module will not
need to know anything about terminal.